+++ docs/about/release-notes.rst
@@ -10,6 +10,7 @@ Bugfixes
 ~~~~~~~~
 
 * Change output dtype of graphs for coordinate transformations to always be floating point, fixes incorrect truncation of the result to integer if, e.g. ``tof`` is an integer (this also affects ``convert``) `#230 <https://github.com/scipp/scippneutron/pull/230>`_.
+* Fix bug in ``load_nexus`` which prevented nexus files containing any empty datasets from being loaded correctly.
 
 v0.4.1 (November 2021)
 ----------------------